--- org/example/Player.java
@@ -1,14 +1,0 @@
-package org.example;
-
-public class Player {
-
-    private final char marker;
-
-    public char getMarker(){
-        return this.marker;
-    }
-
-    Player(char marker){
-        this.marker = marker;
-    }
-}
