--- org/example/PlayfieldsExceptions/PlayfieldsCharInputException.java
@@ -1,10 +1,0 @@
-package org.example.PlayfieldsExceptions;
-
-import org.example.PlayfieldsException;
-
-public class PlayfieldsCharInputException extends PlayfieldsException {
-
-    public PlayfieldsCharInputException(String message) {
-        super(message);
-    }
-}
