--- org/example/PlayfieldsExceptions/PlayfieldsClearFailed.java
@@ -1,9 +1,0 @@
-package org.example.PlayfieldsExceptions;
-
-import org.example.PlayfieldsException;
-
-public class PlayfieldsClearFailed extends PlayfieldsException {
-    public PlayfieldsClearFailed(String message){
-        super(message);
-    }
-}
