--- org/example/PlayfieldsExceptions/PlayfieldsDimensionException.java
@@ -1,10 +1,0 @@
-package org.example.PlayfieldsExceptions;
-
-import org.example.PlayfieldsException;
-
-public class PlayfieldsDimensionException extends PlayfieldsException {
-
-    public PlayfieldsDimensionException(String message){
-        super(message);
-    }
-}
