--- org/example/PlayfieldsExceptions/PlayfieldsSetOccupiedException.java
@@ -1,10 +1,0 @@
-package org.example.PlayfieldsExceptions;
-
-import org.example.PlayfieldsException;
-
-public class PlayfieldsSetOccupiedException extends PlayfieldsException {
-
-    public PlayfieldsSetOccupiedException(String message){
-        super(message);
-    }
-}
