--- src/main/java/org/example/PlayfieldsException.java
@@ -1,11 +1,0 @@
-package org.example;
-
-public class PlayfieldsException extends Exception {
-
-
-    public PlayfieldsException(String message) {
-        super(message);
-    }
-
-
-}
